Mesh

+++ shader/shader.vert
@@ -5,9 +5,9 @@
 
 layout(location = 0) out vec4	out_color;
 
-layout(location = 0) in vec3	in_pos;
-layout(location = 1) in vec3	in_color;
-layout(location = 2) in int		in_ao;
+layout(location = 0) in vec3			in_pos;
+layout(location = 1) in vec3			in_color;
+layout(location = 2) in lowp int		in_ao;
 
 layout(binding = 0) uniform UniformBufferObject {
 	mat4	model;
